# docs/openapi.yaml
openapi: 3.0.3
info:
  title: shareMall API
  description: 移动端 H5 商城（贡献金 + 数字藏品 + 提现）接口规范
  version: 1.0.0
servers:
  - url: https://api.sharemall.com
    description: 生产
  - url: http://localhost:3000
    description: 本地
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/user/kyc:
    post:
      tags: [User]
      summary: 提交实名认证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [realName, idCardNo]
              properties:
                realName: { type: string }
                idCardNo: { type: string }
      responses:
        '200': { description: 返回认证状态 }
...
  /api/nft/{id}/price-history:
    get:
      tags: [NFT]
      summary: 藏品价格走势
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: days, in: query, schema: { type: integer, default: 30 } }
      responses:
        '200': { description: '[{ date, price, changePct }]' }
  /api/nft/{id}:
    get:
      tags: [NFT]
      summary: 藏品详情（含价格走势）
      security: []
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        '200': { description: 含 priceHistory, dealPriceMin/Max }
  /api/nft/{id}/exchange:
    post:
      tags: [NFT]
      summary: 兑换藏品（动态成交价）
      description: dealPrice = currentPrice × (1 + random × dealPremiumPct)，扣 available_fund
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        '200': { description: 返回 dealPrice, referencePrice, serialNo }
...
  /api/nft/listings:
    post:
      tags: [NFT]
      summary: 挂单卖出（按当前参考价，无需传 price）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userNftId]
              properties:
                userNftId: { type: integer }
      responses:
        '200': { description: 返回 referencePrice, dealPriceMin/Max, 预计到账 }
  /api/nft/listings/{id}:
    put:
      tags: [NFT]
      summary: 同步挂单参考价（同步当前价）
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        '200': { description: OK }
    delete:
      tags: [NFT]
      summary: 撤单
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        '200': { description: OK }
...
  /api/nft/trade/{listingId}/buy:
    post:
      tags: [NFT]
      summary: 购买挂单（动态成交价）
      description: |
        dealPrice = currentPrice × (1 + random × dealPremiumPct)。
        买家扣 available_fund（nft_trade_buy）；卖家 withdrawable_cash +(price-fee)，不写 fund_record。
      parameters: [{ name: listingId, in: path, required: true, schema: { type: integer } }]
      responses:
        '200': { description: 返回 price(dealPrice), referencePrice, fee, sellerIncome }
...
  /api/user/profile:
    get:
      tags: [User]
      summary: 个人信息
      responses:
        '200': { description: OK }
    put:
      tags: [User]
      summary: 更新个人信息
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname: { type: string }
                avatar: { type: string }
                gender: { type: string, enum: [unknown, male, female] }
                birthday: { type: string, format: date }
      responses:
        '200': { description: OK }
  /api/user/address:
    get:
      tags: [User]
      summary: 地址列表
      responses:
        '200': { description: OK }
    post:
      tags: [User]
      summary: 新增地址
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [receiver, phone, province, city, district, detail]
              properties:
                receiver: { type: string }
                phone: { type: string }
                province: { type: string }
                city: { type: string }
                district: { type: string }
                detail: { type: string }
                isDefault: { type: boolean }
      responses:
        '200': { description: OK }
...
  /api/withdraw:
    post:
      tags: [User]
      summary: 申请提现（提现金）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, method, accountInfo]
              properties:
                amount: { type: number }
                method: { type: string, enum: [bank, wechat, alipay] }
                accountInfo: { type: object }
      responses:
        '200': { description: 校验提现金余额与实名后受理 }
  /api/withdraw/records:
    get:
      tags: [User]
      summary: 提现记录
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
      responses:
        '200': { description: 分页 }